requirements.txt
django
django-allauth
PyJWT
cryptography
Pillow
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
python-dotenv
django-cors-headers
djangorestframework
requests